Create TC-Calc
The code takes in a destination city, determining a preset Airfare price. This could be improved by creating a dictionary with 'CITYPrice' as the index and the numerical cost as the value. Another dictionary using spelt out city names along with a respective 'CITYPrice' could be added to account for the varying input types.

Number of hotel nights, rental car days, and spending money is also taken. Hotel and car days are multiplied by a preset modifier, though this could be improved by adding a list of average hotel/car rental values for an input city.

The total values are all calculated together, creating total travel expenses.